+++ lib/pitchfork.rb
@@ -26,9 +26,10 @@ module Pitchfork
   # application dispatch.  This is always raised with an empty backtrace
   # since there is nothing in the application stack that is responsible
   # for client shutdowns/disconnects.  This exception is visible to Rack
-  # applications unless PrereadInput middleware is loaded.  This
-  # is a subclass of the standard EOFError class and applications should
-  # not rescue it explicitly, but rescue EOFError instead.
+  # applications.  This is a subclass of the standard EOFError class and
+  # applications should not rescue it explicitly, but rescue EOFError instead.
+  # Such an error is likely an indication that the reverse proxy in front
+  # of Pitchfork isn't properly buffering requests.
   ClientShutdown = Class.new(EOFError)
 
   BootFailure = Class.new(StandardError)