+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Unreleased
 
+# 0.6.0
+
 - Expose `Pitchfork::Info.workers_count` and `.live_workers_count` to be consumed by application health checks.
 - Implement `before_worker_exit` callback.
 - Make each mold and worker a process group leader.